+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Instrument presets. The presets are embedded in the executable and
   there's a button to open a menu to load one of the presets.
 - Frequency modulation target for oscillator, as it was in 4klang
+- Reverb preset settings for a delay unit, with stereo, left and right
+  options
 
 ### Fixed
 - The sointu-vsti-native plugin has different plugin ID and plugin name